requirements.txt
setuptools
wheel
Kekik
PyYAML

Flask[async]
Flask_Sitemap
Flask-WTF
waitress
hypercorn
asgiref
user_agents

asyncio
requests
httpx
aiocached


regex
parsel